+++ SomerenModel/DranklijstItem.cs
@@ -15,7 +15,7 @@ namespace SomerenModel
         public DateTime tijd { get; set; }
         public int drankID { get; set; }
         public int TafelID { get; set; }
-        public float Prijs { get; set; }
+        public double Prijs { get; set; }
         public Alcholisch alcoholisch { get; set; }
     }
 }
